+++ src/main/java/com/example/hotdeal/domain/user/auth/domain/request/SignupRequest.java
@@ -1,4 +1,6 @@
-package com.example.hotdeal.domain.user.auth.domain;
+package com.example.hotdeal.domain.user.auth.domain.request;
+
+import com.example.hotdeal.domain.user.auth.domain.Auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -10,11 +12,11 @@ public class SignupRequest {
     @Email @NotBlank
     private String email;
     @NotBlank
-    private String username;
+    private String name;
     @NotBlank
     private String password;
 
     public Auth toAuth() {
-        return new Auth(this.email, this.username, this.password);
+        return new Auth(this.email, this.name, this.password);
     }
 }
